requirements.txt
django>=4.2,<5.0
djangorestframework>=3.16
gunicorn>=23.0
whitenoise>=6.6
dj-database-url>=3.0
psycopg2-binary>=2.9
requests>=2.32
python-dotenv>=1.0


